Privatised `DefiningString()` and moved it into the "Hash.cs" file
