+++ BHoM_Engine/BHoM_Engine.csproj
@@ -61,7 +61,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Compute\Hash.cs" />
-    <Compile Include="Compute\DefiningString.cs" />
     <Compile Include="Create\BHoMGroup.cs" />
     <Compile Include="Create\IObject\RandomObject.cs" />
     <Compile Include="Create\CustomObject.cs" />
